Dismantle iterrows loop on track DataFrame
Take advantage of fact that it is only one row now.
Convert to Series with .squeeze

Counter for progress through stormlist.

Use str.contains on storm reports to filter event types.

Try to shelter from ZeroDivision error in xarray plot contour. Maybe
something to do with contour levels < 2 or <1 or colorbar, or colors,
or norm=None. . .